+++ openlogin-common/src/main/java/com/nickuc/openlogin/common/settings/Messages.java
@@ -8,6 +8,7 @@
 package com.nickuc.openlogin.common.settings;
 
 import com.nickuc.openlogin.common.model.Title;
+import com.nickuc.openlogin.common.utils.ChatColor;
 import lombok.Getter;
 import lombok.NonNull;
 
@@ -71,38 +72,14 @@ public enum Messages {
      */
     public static void define(@NonNull Messages message, Object value) {
         if (value instanceof String) {
-            value = translateAlternateColorCodes('&', (String) value);
+            value = ChatColor.translateAlternateColorCodes('&', (String) value);
         } else if (value instanceof List) {
             List<Object> list = (List<Object>) value;
             if (!list.isEmpty() && list.get(0) instanceof String) {
-                list.replaceAll(a -> translateAlternateColorCodes('&', (String) a));
+                list.replaceAll(a -> ChatColor.translateAlternateColorCodes('&', (String) a));
             }
         }
         Settings.SETTINGS.put(message.key, value);
-    }
-
-    private static final String CHARACTERS = "0123456789AaBbCcDdEeFfKkLlMmNnOoRr";
-    private static final char COLOR_CODE = 167;
-
-    /**
-     * Translates a string using an alternate color code character into a
-     * string that uses the internal ChatColor.COLOR_CODE color code
-     * character. The alternate color code character will only be replaced if
-     * it is immediately followed by 0-9, A-F, a-f, K-O, k-o, R or r.
-     *
-     * @param altColorChar The alternate color code character to replace. Ex: {@literal &}
-     * @param textToTranslate Text containing the alternate color code character.
-     * @return Text containing the Messages.COLOR_CODE color code character.
-     */
-    public static String translateAlternateColorCodes(char altColorChar, @NonNull String textToTranslate) {
-        char[] b = textToTranslate.toCharArray();
-        for (int i = 0; i < b.length - 1; i++) {
-            if (b[i] == altColorChar && CHARACTERS.indexOf(b[i+1]) > -1) {
-                b[i] = COLOR_CODE;
-                b[i+1] = Character.toLowerCase(b[i+1]);
-            }
-        }
-        return new String(b);
     }
 
     public String asString() {
